Backend cleanup [AK-60] (#18)
* First glance

* Replace console log with logger & started adding more tests

* First cleanup

* Implement `updateTask`

* Change prettier config & format project

* Format files

* Final markups

* Add `composite: true` to tsconfig

* Remove unused function

+++ apps/server/src/tasks/tasks.resolver.ts
@@ -1,9 +1,10 @@
-import { Args, Context, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { ContextRequest } from '../types/context.type';
+import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { CurrentUser } from '../providers/user.provider';
+import { User } from '../users/entities/user.entity';
 import { CreateConstTaskInput } from './dto/createConstTask.input';
 import { CreateFloatTaskInput } from './dto/createFloatTask.input';
 import { GetTasksInput } from './dto/getTasks.input';
-import { UpdateTaskInput } from './dto/update-task.input';
+import { TaskInput } from './dto/task.input';
 import { Task } from './entities/task.entity';
 import { TasksService } from './tasks.service';
 
@@ -13,46 +14,60 @@ export class TasksResolver {
 
   @Mutation(() => Task)
   addConstTask(
-    @Context('req') contextRequest: ContextRequest,
+    @CurrentUser() user: User,
     @Args('createConstTaskInput') CreateConstTaskInput: CreateConstTaskInput
   ) {
-    return this.tasksService.createConst(
-      contextRequest.user,
-      CreateConstTaskInput
-    );
+    return this.tasksService.createConst(user, CreateConstTaskInput);
   }
 
   @Mutation(() => Task)
   addFloatTask(
-    @Context('req') contextRequest: ContextRequest,
+    @CurrentUser() user: User,
     @Args('createFloatTaskInput') createFloatTaskInput: CreateFloatTaskInput
   ) {
-    return this.tasksService.createFloatTask(
-      contextRequest.user,
-      createFloatTaskInput
-    );
+    return this.tasksService.createFloatTask(user, createFloatTaskInput);
   }
 
   @Query(() => [Task])
   getTasks(
-    @Context('req') contextRequest: ContextRequest,
+    @CurrentUser() user: User,
     @Args('getTasksInput') getTasksInput: GetTasksInput
   ) {
-    return this.tasksService.getTasks(contextRequest.user, getTasksInput);
-  }
-
-  @Query(() => Task)
-  findOne(@Args('id', { type: () => Int }) id: number) {
-    return this.tasksService.findOne(id);
+    return this.tasksService.getTasks(user, getTasksInput);
   }
 
   @Mutation(() => Task)
-  updateTask(@Args('updateTaskInput') updateTaskInput: UpdateTaskInput) {
-    return this.tasksService.update(updateTaskInput.id, updateTaskInput);
+  async updateTask(
+    @CurrentUser() user: User,
+    @Args('taskInput') taskInput: TaskInput
+  ) {
+    const task = await Task.findOne(taskInput.id);
+
+    if (!task) {
+      throw new Error('Task not found');
+    }
+    if ((await task.user).id !== user?.id) {
+      throw new Error('Task does not belong to user');
+    }
+
+    const result = await this.tasksService.update(taskInput);
+    return result;
   }
 
   @Mutation(() => Task)
-  removeTask(@Args('id', { type: () => Int }) id: number) {
+  async removeTask(
+    @CurrentUser() user: User,
+    @Args('id', { type: () => Int }) id: number
+  ) {
+    const task = await Task.findOne(id);
+
+    if (!task) {
+      throw new Error('Task not found');
+    }
+    if ((await task.user).id !== user?.id) {
+      throw new Error('Task does not belong to user');
+    }
+
     return this.tasksService.remove(id);
   }
 }
